requirements.txt
fastapi
uvicorn
pillow
pytesseract
python-multipart
opencv-python
pillow  